+++ software/deimos/src/calc/pid.rs
@@ -5,7 +5,7 @@ use crate::{calc_config, calc_input_names, calc_output_names};
 
 /// A PID controller with simple saturation for anti-windup
 #[cfg_attr(feature = "ser", derive(Serialize, Deserialize))]
-#[derive(Default)]
+#[derive(Default, Debug)]
 pub struct Pid {
     // User inputs
     measurement_name: String,