+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2025-05-06 deimos 0.9.0, deimos_gui 0.1.0
+
+### Changed
+
+* !Add bound on Debug for Calc and Peripheral to support use as part of state objects in iced gui
+
 ## 2025-05-03 deimos 0.8.1
 
 Edits to prep for implementation of a calc node editor GUI.